backend/requirements-pip.txt
# Minimal backend Python dependencies for BrainTease
Django
djangorestframework
firebase-admin

# Optional / recommended:
psycopg2-binary  # if you plan to use Postgres in production
pytest
pytest-django
